pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = ""
authors = [
    {name = "Neil Gordon Clark",email = "neilgordonclark@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "pydantic>=2.7.4,<2.10.0",
    "sqlalchemy>=2.0.25,<3.0.0",
    "alembic>=1.13.3,<2.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "pgvector>=0.2.3,<0.3.0",
    "requests>=2.32.3,<3.0.0",
    "openai>=1.60.0,<2.0.0",
    "anthropic>=0.43.0,<0.44.0",
    "tiktoken>=0.9.0,<0.10.0",
    "numpy>=1.26.2,<2.0.0",
    "transformers (==4.51.3)",
    "sentence-transformers>=3.4.1,<4.0.0",
    "letta @ file:///Users/pythagor/nexus/letta",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "accelerate (==1.5.1)",
    "huggingface-hub (==0.30.2)",
    "datasets (>=3.5.0,<4.0.0)",
    "einops (>=0.8.1,<0.9.0)",
    "torch (>=2.6.0,<3.0.0)",
    "sentencepiece (==0.2.0)",
    "chardet>=5.0.0,<6.0.0",
    "python-docx>=1.1.0,<2.0.0",
    "llama-cpp-python>=0.3.0,<0.4.0"
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0,<26.0.0",
    "pytest>=8.3.5,<9.0.0",
    "mypy>=1.15.0,<2.0.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
letta = {develop = true}
torch = {allow-prereleases = true}

[tool.poetry.group.dev.dependencies]
huggingface-hub = "^0.30.2"

